Move _pydict_global_version to _PyRuntimeState.
